Add infrastructure for ternary relation. Add custom cascading delete

+++ backend/src/main/java/ua/yatsergray/backend/service/band/impl/BandAccessRoleServiceImpl.java
@@ -1,14 +1,17 @@
 package ua.yatsergray.backend.service.band.impl;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.yatsergray.backend.domain.dto.band.BandAccessRoleDTO;
 import ua.yatsergray.backend.domain.dto.band.editable.BandAccessRoleEditableDTO;
 import ua.yatsergray.backend.domain.entity.band.BandAccessRole;
+import ua.yatsergray.backend.exception.ChildEntityExistsException;
 import ua.yatsergray.backend.exception.band.BandAccessRoleAlreadyExistsException;
 import ua.yatsergray.backend.exception.band.NoSuchBandAccessRoleException;
 import ua.yatsergray.backend.mapper.band.BandAccessRoleMapper;
 import ua.yatsergray.backend.repository.band.BandAccessRoleRepository;
+import ua.yatsergray.backend.repository.band.BandUserAccessRoleRepository;
 import ua.yatsergray.backend.service.band.BandAccessRoleService;
 
 import java.util.List;
@@ -16,45 +19,48 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
+@Transactional
 @Service
 public class BandAccessRoleServiceImpl implements BandAccessRoleService {
-    private final BandAccessRoleMapper bandAccessRoleMapper;
     private final BandAccessRoleRepository bandAccessRoleRepository;
+    private final BandUserAccessRoleRepository bandUserAccessRoleRepository;
 
     @Autowired
-    public BandAccessRoleServiceImpl(BandAccessRoleMapper bandAccessRoleMapper, BandAccessRoleRepository bandAccessRoleRepository) {
-        this.bandAccessRoleMapper = bandAccessRoleMapper;
+    public BandAccessRoleServiceImpl(BandAccessRoleRepository bandAccessRoleRepository, BandUserAccessRoleRepository bandUserAccessRoleRepository) {
         this.bandAccessRoleRepository = bandAccessRoleRepository;
+        this.bandUserAccessRoleRepository = bandUserAccessRoleRepository;
     }
 
     @Override
     public BandAccessRoleDTO addBandAccessRole(BandAccessRoleEditableDTO bandAccessRoleEditableDTO) throws BandAccessRoleAlreadyExistsException {
-        return bandAccessRoleMapper.mapToBandAccessRoleDTO(bandAccessRoleRepository.save(configureBandAccessRole(new BandAccessRole(), bandAccessRoleEditableDTO)));
+        return BandAccessRoleMapper.INSTANCE.mapToBandAccessRoleDTO(bandAccessRoleRepository.save(configureBandAccessRole(new BandAccessRole(), bandAccessRoleEditableDTO)));
     }
 
     @Override
     public Optional<BandAccessRoleDTO> getBandAccessRoleById(UUID bandAccessRoleId) {
-        return bandAccessRoleRepository.findById(bandAccessRoleId).map(bandAccessRoleMapper::mapToBandAccessRoleDTO);
+        return bandAccessRoleRepository.findById(bandAccessRoleId).map(BandAccessRoleMapper.INSTANCE::mapToBandAccessRoleDTO);
     }
 
     @Override
     public List<BandAccessRoleDTO> getAllBandAccessRoles() {
-        return bandAccessRoleMapper.mapAllToBandAccessRoleDTOList(bandAccessRoleRepository.findAll());
+        return BandAccessRoleMapper.INSTANCE.mapAllToBandAccessRoleDTOList(bandAccessRoleRepository.findAll());
     }
 
     @Override
     public BandAccessRoleDTO modifyBandAccessRoleById(UUID bandAccessRoleId, BandAccessRoleEditableDTO bandAccessRoleEditableDTO) throws NoSuchBandAccessRoleException, BandAccessRoleAlreadyExistsException {
         BandAccessRole bandAccessRole = bandAccessRoleRepository.findById(bandAccessRoleId)
-                .orElseThrow(() -> new NoSuchBandAccessRoleException(String.format("Band access role with id=%s does not exist", bandAccessRoleId)));
+                .orElseThrow(() -> new NoSuchBandAccessRoleException(String.format("Band access role with id=\"%s\" does not exist", bandAccessRoleId)));
 
-        return bandAccessRoleMapper.mapToBandAccessRoleDTO(bandAccessRoleRepository.save(configureBandAccessRole(bandAccessRole, bandAccessRoleEditableDTO)));
+        return BandAccessRoleMapper.INSTANCE.mapToBandAccessRoleDTO(bandAccessRoleRepository.save(configureBandAccessRole(bandAccessRole, bandAccessRoleEditableDTO)));
     }
 
     @Override
-    public void removeBandAccessRoleById(UUID bandAccessRoleId) throws NoSuchBandAccessRoleException {
+    public void removeBandAccessRoleById(UUID bandAccessRoleId) throws NoSuchBandAccessRoleException, ChildEntityExistsException {
         if (!bandAccessRoleRepository.existsById(bandAccessRoleId)) {
-            throw new NoSuchBandAccessRoleException(String.format("Band access role with id=%s does not exist", bandAccessRoleId));
+            throw new NoSuchBandAccessRoleException(String.format("Band access role with id=\"%s\" does not exist", bandAccessRoleId));
         }
+
+        checkIfBandAccessRoleHasChildEntity(bandAccessRoleId);
 
         bandAccessRoleRepository.deleteById(bandAccessRoleId);
     }
@@ -62,19 +68,19 @@ public class BandAccessRoleServiceImpl implements BandAccessRoleService {
     private BandAccessRole configureBandAccessRole(BandAccessRole bandAccessRole, BandAccessRoleEditableDTO bandAccessRoleEditableDTO) throws BandAccessRoleAlreadyExistsException {
         if (Objects.isNull(bandAccessRole.getId())) {
             if (bandAccessRoleRepository.existsByName(bandAccessRoleEditableDTO.getName())) {
-                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with name=%s already exists", bandAccessRoleEditableDTO.getName()));
+                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with name=\"%s\" already exists", bandAccessRoleEditableDTO.getName()));
             }
 
             if (bandAccessRoleRepository.existsByType(bandAccessRoleEditableDTO.getType())) {
-                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with type=%s already exists", bandAccessRoleEditableDTO.getType()));
+                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with type=\"%s\" already exists", bandAccessRoleEditableDTO.getType()));
             }
         } else {
             if (!bandAccessRoleEditableDTO.getName().equals(bandAccessRole.getName()) && bandAccessRoleRepository.existsByName(bandAccessRoleEditableDTO.getName())) {
-                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with name=%s already exists", bandAccessRoleEditableDTO.getName()));
+                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with name=\"%s\" already exists", bandAccessRoleEditableDTO.getName()));
             }
 
             if (!bandAccessRoleEditableDTO.getType().equals(bandAccessRole.getType()) && bandAccessRoleRepository.existsByType(bandAccessRoleEditableDTO.getType())) {
-                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with type=%s already exists", bandAccessRoleEditableDTO.getType()));
+                throw new BandAccessRoleAlreadyExistsException(String.format("Band access role with type=\"%s\" already exists", bandAccessRoleEditableDTO.getType()));
             }
         }
 
@@ -82,5 +88,13 @@ public class BandAccessRoleServiceImpl implements BandAccessRoleService {
         bandAccessRole.setType(bandAccessRoleEditableDTO.getType());
 
         return bandAccessRole;
+    }
+
+    private void checkIfBandAccessRoleHasChildEntity(UUID bandAccessRoleId) throws ChildEntityExistsException {
+        long bandAccessRoleChildEntityAmount = bandUserAccessRoleRepository.countByBandAccessRoleId(bandAccessRoleId);
+
+        if (bandAccessRoleChildEntityAmount > 0) {
+            throw new ChildEntityExistsException(String.format("%s Band user access role(s) depend(s) on the Band access role with id=%s", bandAccessRoleChildEntityAmount, bandAccessRoleId));
+        }
     }
 }
