+++ backend/src/main/java/ua/yatsergray/backend/service/band/impl/ChatAccessRoleServiceImpl.java
@@ -1,14 +1,17 @@
 package ua.yatsergray.backend.service.band.impl;
 
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ua.yatsergray.backend.domain.dto.band.ChatAccessRoleDTO;
 import ua.yatsergray.backend.domain.dto.band.editable.ChatAccessRoleEditableDTO;
 import ua.yatsergray.backend.domain.entity.band.ChatAccessRole;
+import ua.yatsergray.backend.exception.ChildEntityExistsException;
 import ua.yatsergray.backend.exception.band.ChatAccessRoleAlreadyExistsException;
 import ua.yatsergray.backend.exception.band.NoSuchChatAccessRoleException;
 import ua.yatsergray.backend.mapper.band.ChatAccessRoleMapper;
 import ua.yatsergray.backend.repository.band.ChatAccessRoleRepository;
+import ua.yatsergray.backend.repository.band.ChatUserAccessRoleRepository;
 import ua.yatsergray.backend.service.band.ChatAccessRoleService;
 
 import java.util.List;
@@ -16,15 +19,18 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
+@Transactional
 @Service
 public class ChatAccessRoleServiceImpl implements ChatAccessRoleService {
     private final ChatAccessRoleMapper chatAccessRoleMapper;
     private final ChatAccessRoleRepository chatAccessRoleRepository;
+    private final ChatUserAccessRoleRepository chatUserAccessRoleRepository;
 
     @Autowired
-    public ChatAccessRoleServiceImpl(ChatAccessRoleMapper chatAccessRoleMapper, ChatAccessRoleRepository chatAccessRoleRepository) {
+    public ChatAccessRoleServiceImpl(ChatAccessRoleMapper chatAccessRoleMapper, ChatAccessRoleRepository chatAccessRoleRepository, ChatUserAccessRoleRepository chatUserAccessRoleRepository) {
         this.chatAccessRoleMapper = chatAccessRoleMapper;
         this.chatAccessRoleRepository = chatAccessRoleRepository;
+        this.chatUserAccessRoleRepository = chatUserAccessRoleRepository;
     }
 
     @Override
@@ -45,16 +51,18 @@ public class ChatAccessRoleServiceImpl implements ChatAccessRoleService {
     @Override
     public ChatAccessRoleDTO modifyChatAccessRoleById(UUID chatAccessRoleId, ChatAccessRoleEditableDTO chatAccessRoleEditableDTO) throws NoSuchChatAccessRoleException, ChatAccessRoleAlreadyExistsException {
         ChatAccessRole chatAccessRole = chatAccessRoleRepository.findById(chatAccessRoleId)
-                .orElseThrow(() -> new NoSuchChatAccessRoleException(String.format("Chat access role with id=%s does not exist", chatAccessRoleId)));
+                .orElseThrow(() -> new NoSuchChatAccessRoleException(String.format("Chat access role with id=\"%s\" does not exist", chatAccessRoleId)));
 
         return chatAccessRoleMapper.mapToChatAccessRoleDTO(chatAccessRoleRepository.save(configureChatAccessRole(chatAccessRole, chatAccessRoleEditableDTO)));
     }
 
     @Override
-    public void removeChatAccessRoleById(UUID chatAccessRoleId) throws NoSuchChatAccessRoleException {
+    public void removeChatAccessRoleById(UUID chatAccessRoleId) throws NoSuchChatAccessRoleException, ChildEntityExistsException {
         if (!chatAccessRoleRepository.existsById(chatAccessRoleId)) {
-            throw new NoSuchChatAccessRoleException(String.format("Chat access role with id=%s does not exist", chatAccessRoleId));
+            throw new NoSuchChatAccessRoleException(String.format("Chat access role with id=\"%s\" does not exist", chatAccessRoleId));
         }
+
+        checkIfChatAccessRoleHasChildEntity(chatAccessRoleId);
 
         chatAccessRoleRepository.deleteById(chatAccessRoleId);
     }
@@ -62,19 +70,19 @@ public class ChatAccessRoleServiceImpl implements ChatAccessRoleService {
     private ChatAccessRole configureChatAccessRole(ChatAccessRole chatAccessRole, ChatAccessRoleEditableDTO chatAccessRoleEditableDTO) throws ChatAccessRoleAlreadyExistsException {
         if (Objects.isNull(chatAccessRole.getId())) {
             if (chatAccessRoleRepository.existsByName(chatAccessRoleEditableDTO.getName())) {
-                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with name=%s already exists", chatAccessRoleEditableDTO.getName()));
+                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with name=\"%s\" already exists", chatAccessRoleEditableDTO.getName()));
             }
 
             if (chatAccessRoleRepository.existsByType(chatAccessRoleEditableDTO.getType())) {
-                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with type=%s already exists", chatAccessRoleEditableDTO.getType()));
+                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with type=\"%s\" already exists", chatAccessRoleEditableDTO.getType()));
             }
         } else {
             if (!chatAccessRoleEditableDTO.getName().equals(chatAccessRole.getName()) && chatAccessRoleRepository.existsByName(chatAccessRoleEditableDTO.getName())) {
-                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with name=%s already exists", chatAccessRoleEditableDTO.getName()));
+                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with name=\"%s\" already exists", chatAccessRoleEditableDTO.getName()));
             }
 
             if (!chatAccessRoleEditableDTO.getType().equals(chatAccessRole.getType()) && chatAccessRoleRepository.existsByType(chatAccessRoleEditableDTO.getType())) {
-                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with type=%s already exists", chatAccessRoleEditableDTO.getType()));
+                throw new ChatAccessRoleAlreadyExistsException(String.format("Chat access role with type=\"%s\" already exists", chatAccessRoleEditableDTO.getType()));
             }
         }
 
@@ -82,5 +90,13 @@ public class ChatAccessRoleServiceImpl implements ChatAccessRoleService {
         chatAccessRole.setType(chatAccessRoleEditableDTO.getType());
 
         return chatAccessRole;
+    }
+
+    private void checkIfChatAccessRoleHasChildEntity(UUID chatAccessRoleId) throws ChildEntityExistsException {
+        long chatAccessRoleChildEntityAmount = chatUserAccessRoleRepository.countByChatAccessRoleId(chatAccessRoleId);
+
+        if (chatAccessRoleChildEntityAmount > 0) {
+            throw new ChildEntityExistsException(String.format("%s Chat user access role(s) depend(s) on the Chat access role with id=%s", chatAccessRoleChildEntityAmount, chatAccessRoleId));
+        }
     }
 }
